[강운덕] [PROFILER-31] RPC 통계 기능 추가를 위해 TraceContext에서 TraceFactory로 기능을 분리해냄.
git-svn-id: http://svn.bds.nhncorp.com/pe/hippo-tomcat-profiler/trunk@4083 84d0f5b1-2673-498c-a247-62c4ff18d310

+++ src/main/java/com/nhn/pinpoint/profiler/monitor/metric/Histogram.java
@@ -9,7 +9,7 @@ import com.nhn.pinpoint.common.ServiceType;
 public interface Histogram extends Metric {
     ServiceType getServiceType();
 
-    void addResponseTime(long millis);
+    void addResponseTime(int millis);
 
     HistogramSnapshot createSnapshot();
 }
